Fix test case
